+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Add Donation page
 - Add FAQ page
 - Add Footer
+- Add Terms and Conditions page